user-backend/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.4.0
gunicorn==19.9.0
Jinja2==2.10.3
psycopg2-binary==2.8.3
six==1.12.0
urllib3==1.25.3
